Nødt til å plassere sett på vent funksjonalitet i egen service pga sirkulær avhengighet
