+++ src/main/resources/db/migration/V11__slett_sett_paa_vent_tabell.sql
@@ -1,0 +1,1 @@
+DROP TABLE sett_paa_vent;
